refactor Settings Row Item in settings screen
