+++ lib/features/settings/presentation/veiw/screens/settings_screen.dart
@@ -3,6 +3,7 @@ import 'package:azkary_app/core/theming/light_theme.dart';
 import 'package:azkary_app/core/utils/colors.dart';
 import 'package:azkary_app/core/utils/strings.dart';
 import 'package:azkary_app/features/settings/presentation/veiw/widgets/appbar_setting.dart';
+import 'package:azkary_app/features/settings/presentation/veiw/widgets/settings_row_item.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -65,52 +66,6 @@ class _SettingsScreenState extends State<SettingsScreen> {
           ),
         ],
       ),
-    );
-  }
-}
-
-class SettingsRowItem extends StatelessWidget {
-  const SettingsRowItem({
-    super.key,
-    required this.title,
-    required this.leading,
-  });
-  final String title;
-  final Widget leading;
-
-  @override
-  Widget build(BuildContext context) {
-    bool isLightTheme =
-        context.read<ThemeCubit>().state.themeData == lightTheme;
-    return Column(
-      children: [
-        Container(
-          height: 60.h,
-          width: double.infinity,
-          padding: EdgeInsets.symmetric(horizontal: 10.w),
-          color: isLightTheme
-              ? const Color.fromARGB(255, 242, 242, 242)
-              : const Color.fromARGB(255, 90, 120, 13),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                title,
-                style: TextStyle(fontSize: 18.sp),
-              ),
-              const Spacer(),
-              leading,
-            ],
-          ),
-        ),
-        Divider(
-          color: isLightTheme
-              ? const Color.fromARGB(255, 186, 186, 186)
-              : const Color.fromARGB(255, 249, 249, 249),
-          thickness: 1,
-          height: 1,
-        ),
-      ],
     );
   }
 }
